pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etphonehome"
version = "0.1.4"
description = "Remote access system for Claude CLI via reverse SSH tunnels"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ET Phone Home Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "paramiko>=3.0.0",
    "pyyaml>=6.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
server = [
    "mcp>=1.0.0",
    "aiofiles>=23.0.0",
    "starlette>=0.35.0",
    "uvicorn>=0.27.0",
    "sse-starlette>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
phonehome = "client.phonehome:main"
etphonehome-server = "server.mcp_server:main"

[tool.setuptools.packages.find]
include = ["client*", "server*", "shared*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
